+++ src/main/java/lotto/domain/reword/Reword.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.reword;
 
 import lotto.exception.InvalidLottoHitCountException;
