+++ src/test/java/lotto/domain/LottoTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
+
+import lotto.domain.lotto.Lotto;
 import lotto.exception.DuplicateLottoNumberException;
 import lotto.exception.InvalidLottoNumberCountException;
 import org.junit.jupiter.api.DisplayName;
