+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -6,6 +6,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
